api/requirements.txt
alembic==1.8.1
amqp==5.1.1
aniso8601==9.0.1
apispec==5.2.2
apispec-webframeworks==0.5.2
arrow==1.2.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.2
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.5
dnspython==2.2.1
email-validator==1.1.2
filelock==3.7.1
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-CLI==0.4.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
graphviz==0.20.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
infinity==1.5
inflection==0.5.1
intervals==0.9.2
isbnlib==3.10.3
iso3166==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
kombu==5.2.4
Mako==1.2.1
manage.py==0.2.10
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
numpy==1.23.1
packaging==21.3
pandas==1.4.3
passlib==1.7.4
pg8000==1.29.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
py==1.11.0
py-moneyed==0.8.0
pycountry==22.3.5
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
sadisplay==0.4.9
schwifty==2020.9.0
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.39
sqlalchemy-mixins==1.5.3
sqlalchemy-repr==0.0.2
SQLAlchemy-Utils==0.38.3
toml==0.10.2
tox==3.25.1
validators==0.20.0
vine==5.0.0
virtualenv==20.16.1
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
zipp==3.8.1
